Remove Problems module
