+++ src/totaldegree.jl
